microservicios/alertas/requirements.txt
Flask
psycopg2-binary
kafka-python
